-Zmiana nazwy klienta na serwer -Dodanie GUI do klienta
